[BOOK-182] refactor: extract SearchResultCell content into reusable BKBookSummaryView
